\name{CatDynData}
\alias{CatDynData}
\title{
Generic CatDyn Data Plotting
}
\description{
Generic function to plot CatDyn data objects.
}
\usage{
CatDynData(x, ...)
}
\arguments{
  \item{x}{
An object of class CatDynData, generated by BioFishpDay.Fk().
}
  \item{\dots}{
Not used
}
}
\details{
This generic function will just be used to give a class to the three last components of the list output by
BioFishpDay.Fk().
}
\value{
A class attribute.
}
\author{
Ruben H. Roa-Ureta
}
\note{
Objects of class CatDynData are plotted for exploration of effort measures and
general season trends.
}
\examples{
#See example in man pages for BioFishpDay.Fk() and plot.CatDynData().
}