\name{CatDynExp}
\alias{CatDynExp}
\title{
Class Attribute of Exploratory Analysis Object
}
\description{
To be used by plot.CatDynExp() to explore initial parameter values before fitting 
models.
}
\usage{
CatDynExp(x, ...)
}
\arguments{
  \item{x}{
A list object coming from internal functions.
}
  \item{\dots}{
Not used.
}
}
\value{
A class attribute.
}
\author{
Ruben H. Roa-Ureta
}
\examples{
#See examples for CatDynFit().
}
\keyword{ ~models }

