\name{gayhake}
\alias{gayhake}
\docType{data}
\title{
Industrial and Artisanal Catch and Effort Data from the Chilean Hake Fishery
}
\description{
Exhaustive recorded operational activity of the \emph{Merluccius gayi} two fleet
fishery off Central Chile in 2006.
}
\usage{data("gayhake")}
\format{
  A data frame with 53 observations on the following 7 variables.
  \describe{
    \item{\code{TimeStep}}{a numeric vector}
    \item{\code{Catch.Ind.kg}}{a numeric vector}
    \item{\code{NShips.Ind}}{a numeric vector}
    \item{\code{Body.Ind.g}}{a numeric vector}
    \item{\code{Catch.Art.kg}}{a numeric vector}
    \item{\code{NShips.Art}}{a numeric vector}
    \item{\code{Body.Art.g}}{a numeric vector}
  }
}
\source{
Under-Secretariat of Fishing, Chilean government.
}
\examples{
data(gayhake)
}
\keyword{datasets}
