\name{gayhakelm}
\alias{gayhakelm}
\docType{data}
\title{
Length and Month data of Chilean hake from Artisanal Fishery
}
\description{
A small sample from hundreds of thousand of records of length by month from
individual fish captured by artisanal fishing boats.
}
\usage{data("gayhakelm")}
\format{
  A data frame with 600 observations on the following 2 variables.
  \describe{
    \item{\code{Month}}{a numeric vector}
    \item{\code{lon.cm}}{a numeric vector}
  }
}
\source{
Under-Secretariat of Fishing, Chilean government.
}
\examples{
data(gayhakelm)
}
\keyword{datasets}
