\name{gayhakelw}
\alias{gayhakelw}
\docType{data}
\title{
Length and Weight data of Chilean hake from Artisanal Fishery
}
\description{
A small sample from hundreds of thousand of records of length and weight of
individual fish captured by artisanal fishing boats.
}
\usage{data("gayhakelw")}
\format{
  A data frame with 250 observations on the following 2 variables.
  \describe{
    \item{\code{Length.cm}}{a numeric vector}
    \item{\code{Weight.kg}}{a numeric vector}
  }
}
\source{
Under-Secretariat of Fishing, Chilean government.
}
\examples{
data(gayhakelw)
}
\keyword{datasets}
