\name{m.hoenig}
\alias{m.hoenig}
\docType{data}
\title{
Marine animals natural mortality and longevity database
}
\description{
A database of natural mortality estimates and longetvity in fish, cetaceans and molluscs
compiled by Prof. John M. Hoenig.
}
\usage{data("m.hoenig")}
\format{
  A data frame with 123 observations on the following 2 variables.
  \describe{
    \item{\code{maxage}}{a numeric vector}
    \item{\code{natmort}}{a numeric vector}
  }
}
\details{
Mortality rates in 1/yr and longevity in yr.
}
\references{
Hoening, J.M. 2005. Empirical use of longevity data to estimate mortality rates.
SEDAR33-RD17. SEDAR, North Charleston, SC. 8 pp.
}
\examples{
data(m.hoenig)
}
\keyword{datasets}
