\name{SeasonDates}
\alias{SeasonDates}
\docType{data}
\title{
Star and End Dates of Loligo Seasons in the Falkland Islands
}
\description{
A data.frame with the dates of start and end of Loligo seasons in the ISO 8601
international standard, yyy-mm-dd
}
\usage{data(SeasonDates)}
\format{
  A data frame with 40 observations on the following 3 variables.
  \describe{
    \item{\code{Season}}{a numeric vector}
    \item{\code{Start.Date}}{a Date}
    \item{\code{End.Date}}{a Date}
  }
}
\source{
http://fis.com/falklandfish/
}
\examples{
data(SeasonDates)
SeasonDates
}
\keyword{datasets}
