\name{twelver}
\alias{twelver}
\docType{data}
\title{
Sport Fishing for Eel Juveniles in Taiwan
}
\description{
Exhaustive recorded operational activity of the \emph{Anguilla japonica} sport
fishery in Taiwan during the 1983-1984 season.
}
\usage{data("twelver")}
\format{
  A data frame with 110 observations on the following 4 variables.
  \describe{
    \item{\code{Date}}{a Date}
    \item{\code{Effort.Hour.Soak}}{a numeric vector}
    \item{\code{Catch.Ind}}{a numeric vector}
    \item{\code{Mbw}}{a numeric vector}
  }
}
\source{
Dpt. of Zoology, National Taiwan University
}
\examples{
data(twelver)
}
\keyword{datasets}
