\name{catdyn}
\alias{catdyn}
\title{
Class Attribute of Numerically Fit CatDyn Model Objects
}
\description{
To be used by CatDynPred() to create model results.
}
\usage{
catdyn(x, ...)
}
\arguments{
  \item{x}{
A list object coming from the optimization wrapper CatDynFit().
}
  \item{\dots}{
Not used.
}
}
\value{
A class attribute.
}
\author{
Ruben H. Roa-Ureta
}
\examples{
#See examples for CatDynFit().
}
\keyword{ ~classes }
