\name{CatDynData}
\alias{CatDynData}
\title{
Class Attribute of CatDyn Data Object
}
\description{
To be used by plot.CatDynData() to examine raw data and CatDynFit() to fit models.
}
\usage{
CatDynData(x, ...)
}
\arguments{
  \item{x}{
A list object coming from as.CatDynData().
}
  \item{\dots}{
Not used.
}
}
\value{
A class attribute.
}
\author{
Ruben H. Roa-Ureta
}
\examples{
#See example for as.CatDynData().
}
\keyword{ ~classes }

