\name{CatDynMod}
\alias{CatDynMod}
\title{
Class Attribute of Model Output Object
}
\description{
To be used by plot.CatDynMod() to graphically examine model results.
}
\usage{
CatDynMod(x, ...)
}
\arguments{
  \item{x}{
A list object coming from CatDyn().
}
  \item{\dots}{
Not used.
}
}
\value{
A class attribute
}
\author{
Ruben H. Roa-Ureta
}
\examples{
#See examples for CatDynFit().
}
\keyword{ ~classes }
