\name{lolgahi}
\alias{lolgahi}
\docType{data}
\title{
Industrial Trawling Data from the Falkland Islands
}
\description{
Exhaustive recorded operational activity of the \emph{Loligo gahi} fleet off the Falkland 
Islands during the summer season of 1990.
}
\usage{data(lolgahi)}
\format{
  A data frame with 120 observations on the following 6 variables.
  \describe{
    \item{\code{obscat.f1}}{a numeric vector}
    \item{\code{obseff2.f1}}{a numeric vector}
    \item{\code{obsmbm.f1}}{a numeric vector}
    \item{\code{obscat.f2}}{a numeric vector}
    \item{\code{obseff.f2}}{a numeric vector}
    \item{\code{obsmbm.f2}}{a numeric vector}
  }
}
\source{
http://fis.com/falklandfish/
}
\references{
Roa-Ureta, R. H. 2012. ICES Journal of Marine Science 69(8):1403-1415.
}
\examples{
data(lolgahi)
}
\keyword{datasets}
